requirements.txt
amqp==2.3.2
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
Django==2.0.7
django-celery==3.2.2
django-celery-beat==1.1.1
django-celery-results==1.0.1
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg2==2.7.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.27.0
pymacaroons==0.13.0
PyNaCl==1.1.2
PyQt5==5.10
pyRFC3339==1.0
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
sip==4.19.7
six==1.11.0
suds-jurko==0.6
sweetify==1.0.3
urllib3==1.22
vine==1.1.4
wadllib==1.3.2
whitenoise==3.3.1
zope.interface==4.3.2
